pdp/pdp/pdp/requirements.txt
flask
mysqlclient
sqlalchemy
flask-sqlalchemy
python-dotenv